requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
Pillow==10.0.0
nltk==3.8.1
lime==0.2.0.1
scikit-learn==1.3.0
tqdm==4.65.0
pyyaml==6.0
wandb==0.15.12
jupyter==1.0.0
notebook==7.0.6
opencv-python==4.8.1.78
seaborn==0.12.2
tensorboard==2.14.0